backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Banco de dados
SQLAlchemy==2.0.30
databases==0.9.0
alembic==1.13.1

# Segurança / Autenticação
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validação e Uploads
pydantic==2.8.0
python-multipart==0.0.9
